requirements.txt
astroid==1.6.3
Django==2.0.5
django-crispy-forms==1.7.2
django-formtools==2.1
docxtpl==0.4.10
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.4
pylint==1.8.4
python-docx==0.8.6
pytz==2018.4
six==1.11.0
wrapt==1.10.11
